crimson/os/seastore/btree: move lba/backref_map_val_t to the common header file
rename btree_range_pin.h to btree_types.h

(cherry picked from commit 8c6d82e0c7d6d27a9544d065bc6b9ccec4c6a3cc)
