--- src/crimson/os/seastore/backref/backref_tree_node.cc
@@ -1,14 +1,0 @@
-// -*- mode:C++; tab-width:8; c-basic-offset:2; indent-tabs-mode:t -*-
-// vim: ts=8 sw=2 smarttab
-
-#include "crimson/os/seastore/backref/backref_tree_node.h"
-
-namespace crimson::os::seastore::backref {
-
-std::ostream& operator<<(std::ostream &out, const backref_map_val_t& val) {
-  return out << "backref_map_val_t("
-	     << val.laddr
-	     << "~" << val.len << ")";
-}
-
-} // namespace crimson::os::seastore::backref
