+++ src/crimson/os/seastore/lba_mapping.h
@@ -4,7 +4,7 @@
 #pragma once
 
 #include "crimson/os/seastore/cached_extent.h"
-#include "crimson/os/seastore/btree/btree_range_pin.h"
+#include "crimson/os/seastore/btree/btree_types.h"
 #include "crimson/os/seastore/lba_manager/btree/lba_btree_node.h"
 #include "crimson/os/seastore/logical_child_node.h"
